feat(PlayToken): Includes ERC777 Standard
