+++ contracts.1/ERC777TokensRecipient.sol
@@ -1,0 +1,21 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * This code has not been reviewed.
+ * Do not use or deploy this code before reviewing it personally first.
+ */
+// solhint-disable-next-line compiler-fixed
+pragma solidity ^0.4.24;
+
+
+interface ERC777TokensRecipient {
+    function tokensReceived(
+        address operator,
+        address from,
+        address to,
+        uint256 amount,
+        bytes data,
+        bytes operatorData
+    ) public;
+}
