+++ contracts/ERC20Token.sol
@@ -1,0 +1,26 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * This code has not been reviewed.
+ * Do not use or deploy this code before reviewing it personally first.
+ */
+// solhint-disable-next-line compiler-fixed
+pragma solidity ^0.4.24;
+
+
+interface ERC20Token {
+    function name() public view returns (string);
+    function symbol() public view returns (string);
+    function decimals() public view returns (uint8);
+    function totalSupply() public view returns (uint256);
+    function balanceOf(address owner) public view returns (uint256);
+    function transfer(address to, uint256 amount) public returns (bool);
+    function transferFrom(address from, address to, uint256 amount) public returns (bool);
+    function approve(address spender, uint256 amount) public returns (bool);
+    function allowance(address owner, address spender) public view returns (uint256);
+
+    // solhint-disable-next-line no-simple-event-func-name
+    event Transfer(address indexed from, address indexed to, uint256 amount);
+    event Approval(address indexed owner, address indexed spender, uint256 amount);
+}
