+++ contracts/ERC777Token.sol
@@ -1,0 +1,42 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * This code has not been reviewed.
+ * Do not use or deploy this code before reviewing it personally first.
+ */
+// solhint-disable-next-line compiler-fixed
+pragma solidity ^0.4.24;
+
+
+interface ERC777Token {
+    function name() public view returns (string);
+    function symbol() public view returns (string);
+    function totalSupply() public view returns (uint256);
+    function balanceOf(address owner) public view returns (uint256);
+    function granularity() public view returns (uint256);
+
+    function defaultOperators() public view returns (address[]);
+    function isOperatorFor(address operator, address tokenHolder) public view returns (bool);
+    function authorizeOperator(address operator) public;
+    function revokeOperator(address operator) public;
+
+    //function send(address to, uint256 amount, bytes data) public;
+    function operatorSend(address from, address to, uint256 amount, bytes data, bytes operatorData) public;
+
+    function burn(uint256 amount, bytes data) public;
+    function operatorBurn(address from, uint256 amount, bytes data, bytes operatorData) public;
+
+    event Sent(
+        address indexed operator,
+        address indexed from,
+        address indexed to,
+        uint256 amount,
+        bytes data,
+        bytes operatorData
+    ); // solhint-disable-next-line separate-by-one-line-in-contract
+    event Minted(address indexed operator, address indexed to, uint256 amount, bytes operatorData);
+    event Burned(address indexed operator, address indexed from, uint256 amount, bytes data, bytes operatorData);
+    event AuthorizedOperator(address indexed operator, address indexed tokenHolder);
+    event RevokedOperator(address indexed operator, address indexed tokenHolder);
+}
