+++ package.json
@@ -34,12 +34,13 @@
     "@digix/tempo": "^0.2.0",
     "lodash": "^4.17.5",
     "run-with-testrpc": "^0.3.0",
-    "truffle": "^4.1.5",
+    "truffle": "^4.1.14",
     "truffle-hdwallet-provider": "0.0.3"
   },
   "dependencies": {
     "@gnosis.pm/util-contracts": "^0.2.0",
     "dotenv": "^6.2.0",
+    "erc777": "^0.1.1",
     "npm-prepublish": "^1.2.3",
     "yargs": "^11.0.0"
   }
